+++ content/en/docs/collector/components/processor.md
@@ -10,8 +10,4 @@ Processors transform, filter, and enrich telemetry data as it flows through the
 pipeline. For more information on how to configure processors, see the
 [Collector configuration documentation](/docs/collector/configuration/#processors).
 
-<!-- BEGIN GENERATED: processor-table SOURCE: scripts/collector-sync -->
-
 {{< collector-component-rows type="processor" >}}
-
-<!-- END GENERATED: processor-table SOURCE: scripts/collector-sync -->
